Remove Go 1.18 build tags
Go 1.19 has been the minimum required version
since Go 1.20 was released.
With the release of Go 1.21, even Go 1.19 is unsupported.
There's no reason specialize for versions before Go 1.18 anymore.

This deletes the Go 1.18 build-tagged files
and merges their contents into their siblings.
